+++ src/cmd/services/get/get-service.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"fmt"
 	cloudclient "github.com/otterize/otterize-cli/src/pkg/cloudclient/restapi"
+	"github.com/otterize/otterize-cli/src/pkg/cloudclient/restapi/cloudapi"
 	"github.com/otterize/otterize-cli/src/pkg/config"
 	"github.com/otterize/otterize-cli/src/pkg/output"
 	"github.com/samber/lo"
@@ -13,39 +14,31 @@ import (
 
 var GetCmd = &cobra.Command{
 	Use:          "get <service-id>",
-	Short:        "Get a service for a given id",
+	Short:        "Get details for a service",
 	Args:         cobra.ExactArgs(1),
-	Long:         ``,
 	SilenceUsage: true,
-	RunE:         getService,
-}
+	RunE: func(_ *cobra.Command, args []string) error {
+		ctxTimeout, cancel := context.WithTimeout(context.Background(), config.DefaultTimeout)
+		defer cancel()
 
-func getService(_ *cobra.Command, args []string) error {
-	ctxTimeout, cancel := context.WithTimeout(context.Background(), config.DefaultTimeout)
-	defer cancel()
+		c, err := cloudclient.NewClientFromToken(viper.GetString(config.OtterizeAPIAddressKey), config.GetAPIToken(ctxTimeout))
+		if err != nil {
+			return err
+		}
 
-	client, err := cloudclient.NewClientFromToken(viper.GetString(config.OtterizeAPIAddressKey), config.GetAPIToken(ctxTimeout))
-	if err != nil {
-		return err
-	}
+		id := args[0]
+		resp, err := c.ServiceQueryWithResponse(ctxTimeout, id)
+		if err != nil {
+			return err
+		}
 
-	id := args[0]
+		service := lo.FromPtr(resp.JSON200)
 
-	resp, err := client.ServiceQueryWithResponse(ctxTimeout, id)
-	if err != nil {
-		return err
-	}
-
-	service := lo.FromPtr(resp.JSON200)
-
-	result, err := output.GetFormattedObject(service)
-	if err != nil {
-		return err
-	}
-	fmt.Println(result)
-	return nil
-}
-
-func init() {
-	GetCmd.Flags().String(config.OutputFormatKey, config.OutputFormatDefault, fmt.Sprintf("output format - %s/%s", config.OutputYaml, config.OutputJson))
+		result, err := output.FormatServices([]cloudapi.Service{service})
+		if err != nil {
+			return err
+		}
+		fmt.Println(result)
+		return nil
+	},
 }
