--- src/cmd/services/list/config.go
@@ -1,5 +1,0 @@
-package list
-
-const EnvironmentIDKey = "env-id"
-const NamespaceIDKey = "namespace-id"
-const ServiceNameKey = "name"
