+++ src/cmd/services/services.go
@@ -10,7 +10,8 @@ import (
 var ServicesCmd = &cobra.Command{
 	Use:     "services",
 	GroupID: groups.ResourcesGroup.ID,
-	Short:   "",
+	Aliases: []string{"service"},
+	Short:   "Service operations via the Otterize API",
 }
 
 func init() {
